fix: remove cell data when user enters an empty string
Before if a cell already had some data and user deleted the corresponding string while editing the cell,
the cell value would not update. Now if a cell already had a value, and the user edited the cell and
left an empty string, the corresponding data in the CellMap will be deleted.